backend/requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bcrypt==3.2.0
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.1
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.86.0
future==0.18.2
greenlet==1.1.1
h11==0.14.0
idna==3.4
passlib==1.7.4
pip==22.2.2
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
PyMySQL==1.0.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.7.2
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.39
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
wheel==0.37.1
wincertstore==0.2
